arquivos do prof

+++ src/br/gov/cesarschool/poo/fidelidade/cartao/entidade/LancamentoExtratoPontuacao.java
@@ -1,12 +1,11 @@
 package br.gov.cesarschool.poo.fidelidade.cartao.entidade;
-
 import java.io.Serializable;
 import java.time.LocalDateTime;
 
-public class LancamentoExtratoPontuacao extends LancamentoExtrato implements Serializable{
-    
-    public LancamentoExtratoPontuacao(long numeroCartao, int quantidadePontos, LocalDateTime dataHoraLancamento) {
-        super(numeroCartao, quantidadePontos, dataHoraLancamento);
-    }
-    
+public class LancamentoExtratoPontuacao extends LancamentoExtrato implements Serializable {
+
+	public LancamentoExtratoPontuacao(long numeroCartao, double quantidadePontos, LocalDateTime dataHoraLancamento) {
+		super(numeroCartao, quantidadePontos, dataHoraLancamento);
+	}
+	
 }
